Issue #1824100 by Wim Leers: 'edit ID's have changed, they now also include language code and view mode. The Create.js branch wasn't updated to use this yet.
A side effect was that the predicate (now consisting of field name, language code and view mode, separated by colons) was regarded by VIE as an invalidly namespaced predicate. Switched to slashes as a separator instead to solve this problem. We should swap colons for slashes everywhere though, also in the DOM.